+++ Objects/Plane.h
@@ -3,7 +3,7 @@
 
 #include "Object.h"
 
-#include <Eigen/Dense>
+#include <Eigen/Geometry>
 using namespace Eigen;
 
 
@@ -13,14 +13,14 @@ class Plane : public Object
         EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
     public:
-        Plane(Material *mat, const Vector3d& p, const Vector3d& n)
+        Plane(Material *mat, const Vector3f& p, const Vector3f& n)
         : Object(mat), p(p), n(n) {}
 
         bool intersect(const Ray& ray, IntersectionData *intersectionData) const override;
 
     private:
-        Vector3d p;
-        Vector3d n;
+        Vector3f p;
+        Vector3f n;
 };
 
 #endif // PLANE_H
