+++ Objects/Sphere.cpp
@@ -2,19 +2,19 @@
 
 bool Sphere::intersect(const Ray& ray, IntersectionData *intersectionData) const
 {
-    Vector3d transformedPos = m_t*m_pos;
+    Vector3f transformedPos = m_t*m_pos;
 
-    double a = ray.d.dot(ray.d);
-    Vector3d oc = ray.o - transformedPos;
-    double b = 2*(ray.d.dot(oc));
-    double c = oc.dot(oc) - m_radius*m_radius;
+    float a = ray.d.dot(ray.d);
+    Vector3f oc = ray.o - transformedPos;
+    float b = 2*(ray.d.dot(oc));
+    float c = oc.dot(oc) - m_radius*m_radius;
 
-    double det = b*b - 4*a*c;
+    float det = b*b - 4*a*c;
     if (det < 0)
         return false;
-    double sqrtdet = sqrt(det);
-    double d1 = (-b - sqrtdet) / (2*a);
-    double d2 = (-b + sqrtdet) / (2*a);
+    float sqrtdet = sqrt(det);
+    float d1 = (-b - sqrtdet) / (2*a);
+    float d2 = (-b + sqrtdet) / (2*a);
 
     if (d2 < DELTA_INTERSECT && d1 < DELTA_INTERSECT)
         return false; // the two points land behind the ray
