+++ Utils/Ray.cpp
@@ -1,6 +1,6 @@
 #include "Ray.h"
 
-double Ray::distanceFrom(const Vector3d& p) const
+float Ray::distanceFrom(const Vector3f& p) const
 {
     return (o - p).cross(d).norm() / d.norm();
 }
